safari-600.7-branch/LayoutTests/accessibility/aria-disabled-propagated-to-children-expected.txt

Foo
Bar
This tests that aria-disabled will be taken from an ancestor if available

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS item1.isEnabled is false
PASS item2.isEnabled is true
PASS tablist.isEnabled is false
PASS successfullyParsed is true

TEST COMPLETE
